better PEP-0008 (trans_text_ ==> trans_text)
PEP-0008 reserves the use of a trailing underscore for names
which conflict with Python keywords, which this doesn't.

The goal is to have a permanent name which is unique across
the gramps namespace, since it is defined (in po/genpot.sh
and po/update_po.py) as a keyword to flag strings to be put
into the translation "catalog" file (gramps.pot).

The new keyword "trans_text" is still unique.  (Also "_T_".)


svn: r22200