Bug 7377 Ver 4.0.2 starts in East Asian Language
Revert "GrampsLocale: Don't set  to 'C' if LANG is en_US"

This reverts commit bdd336f38e900a6a2e14c7cd441fba5aad733bec.

It turns out that Gtk's having an 'en' translation isn't the only reason
for doing this: GtkBuilder uses $LANGUAGE too, and will look for an english
translation of strings in our Glade files. Since there's no en.po or
en_US.po, it will go to the next language in the list and produce confusing
results.