Remove gramps.gen.utils.file.fix_encoding()
Yet another overly complicated way to spell conv_to_unicode().
Note that unicode() and str.decode() are alternate spellings of the same function in Py2, and unicode();
note as well that this was used on strings coming from the GUI or the database, not the shell, so that
using glocale.encoding was also incorrect.
(cherry picked from commit f0ffd6e9b9dee138e5c8eca8d418d96f34438b88)