9443: Gramps not appearing in Gnome Software
I updated the appdata.xml file to the latest version of the specification
(https://www.freedesktop.org/software/appstream/docs/index.html), adding
some new tags in the process.

I also installed the appdata.xml file to the latest location in the
specification.

The XML has bene validated by the appstreamcli validation tool, and the
patch applied in Debian (Gramps 4.2.3).

If it looks OK, it should probably be cherry-picked to the gramps42
maintenance branch.