enhance Book class (to allow non-identical Book copies to be made)
Historically, a new Book made from an old Book argument would
contain the exact same BookItem objects as the original Book.
That behavior is retained (even though I can't see that it is
ever used), but now copies are possible where the BookItems
are of the same number and type as the original, but not the
exact same BookItem objects.  (No "paper" options are copied.)

In the process of doing that, the "save" method was slightly
tweaked, so its recently-changed (for pylint reasons) strings
have their XML terminator on the same (Python) line as the
XML-start of the keyword, for more clarity.  In the process of
checking that a pathological case was noticed where the metric
paper option could be Boolean in some cases (instead of the
expected zero/one), so that was coped with also.