Enhance Book Report
The main change is that now output-related options will also
be saved to (and restored from) the books.xml file.  If the
user has three books for instance, each will keep the last
filename it was written to, and present that filename to the
user in the final dialog, for acceptance or modification.
Similarly for things like paper size and orientation.

Also, the error reporting mechanism was enhanced, so that
now if a report inside a book has a ReportError the user
will be told gracefully (instead of gramps crashing).  The
(translated) name of the report will also be shown.

A few more dialogs were given a "parent" also.