move comment lines higher, to keep out of gramps.pot
The problem is that "nearby" comments are put into the
gramps.pot template, to aid the translator, but these
comments are referring to an untranslated string, which
then confuses the translator, who wonders if the string
flagged for translation should be translated after all.

The solution is to move the comment futher away, so it
is no longer "nearby" the translated string.

http://gramps.1791082.n4.nabble.com/Fw-Translation-question-td4676138.html