8128: GtkDialog mapped without a transient parent
In the process of searching for uses of ErrorDialog which
did not have a "parent" argument, I stumbled on this module
-- and noticed something different in it.

Benny changed the original "ErrorDialog" calls to a
"ERROR = ErrorDialog" variant in
https://github.com/gramps-project/gramps/commit/9a2fd506425756531d2bb725d87ad4feefa99986
on 30 January 2011, "to allow a QML interface" (which by the
way has subsequently been removed from the mainline gramps
code).

In the same commit, in cli/clidbman.py, he added an
_errordialog function and an "ERROR = _errordialog" line,
evidently so all DB errors, no matter whether in GUI or CLI
code, could be changed to things like:

ERROR(_("Could not rename family tree"), str(msg))

But then in 16 February 2011 MDNauta, in
https://github.com/gramps-project/gramps/commit/7c1b26d969900d5cf70f4408b9078aeed1d70fcd
changed the generic ERROR to either DbManager.ERROR or else
CLIDbManager.ERROR, depending on whether it was in GUI code
or CLI code.

Which of course defeated the purpose of the change to the
generic call, so I have changed it back to the original
ErrorDialog call, directly.

The decision was made even easier by noticing lots of
more-recent error-handling code which used the ErrorDialog
form as opposed to Benny's variant.