bug 9877; fix performace with people tree view and some filters
When using deep relationship filters on the person tree view with a
large tree, the tree redraw can take a long time (hours).  The
DeepRelationshipPathBetween 'prepare' method was getting called for
every person in the preceding filter. And that it also ran through
every person in the db within the prepare method.  Resulting in time
as a square function of number of people.

Change view code so that 'prepare' method is called once, and 'apply'
method is called once with list of handles to scan, rather than both
called once per handle.