bug 9854; fix Merge Family; Error dialog hidden behind Merge Dialog
While performing a Merge Family that results in an error, the Error
dialog appears as normally initially.

However, if the user switches to another application and back, the
Error dialog can disappear (it is under the Merge dialog in the UI).
The mouse cannot clear this up (although the keyboard 'enter' or
'Esc' keys will be directed to the Error dialog and close everything.