Fixes ManagedWindow for a modal window with another modal elsewhere
During Gramps startup, it is possible to start two modal dialogs;
the 'Family Trees' and the 'Update Addons' (when preferences is set
to check on startup).  Both are modal.
This upgrades the Window manager and ManagedWindow to support this
case.  If this occurs, then the most recent modal window is left
that way, a previous modal window is temporarily downgraded to
non-modal, and the newest window is set 'transient' for the previous
modal window, even if it was not the parent.
This means that there can be only one modal window, and it is always
on top and active.