bug 9927; fix UndoableEntry for pygobject bug
This is a workaround for
https://bugzilla.gnome.org/show_bug.cgi?id=644927
What is happening, is that our UndoableEntry class is connecting
to the 'insert-text' event, which uses one of the 'in-out'
parameters for the position of the cursor in the buffer. The very
fact of the connect to the event causes the issue, not something
Gramps is doing wrong.