graphviz reports: connect arrows to side of nodes instead of anywhere
This sets the headport and tailport attributes to all edges in the
graphviz file. Their default value makes the arrows between persons
and/or family nodes attach their ends to the respective node at any
position. With this change, the position facing the node on the other
side of the arrow is chosen.