Remove duplicated keywords.
The keywords for the name formats have
been duplicated in order to provide a
translated version.

To remove this duplication, the
translation is produced by mapping
over the keywords that was duplicated.

The actual mapping is delayed
until a choose is made on which
should be the new name format hereby
reducing the amount of translation that
is actually done.