Remove the resource-path file
Attempt to derive the resource path from the package path assuming that
one of the three main installation schemes has been used.

The package path will be one of the following:

  <prefix>/lib/pythonX.Y/site-packages
  <prefix>\Lib\site-packages

  <home>/lib/python

  <userbase>/lib/pythonX.Y/site-packages
  <userbase>\PythonXY\site-packages

  Where <prefix>, <home> and <userbase> are the resource paths used in the
  Prefix, Home and User installation schemes.

The use of the command line option "--install-data" in the setup script
is no longer supported.

This change is intended to allow a Gramps core package to be created
as a wheel and installed via pip.