Install MIME icons in global hicolor icon theme
At the moment, the icons of MIME types are installed in a location
specific for GNOME, which is not usable when not using a GNOME icon
theme. Also, the naming is a deprecated GNOME one, that might not be
even working anymore.

Instead, rename the icons according to the MIME they represent, and
install them in the global hicolor icon theme: this way, they are used
with any icon theme that does not provide own versions of them (which
is the vast majority of the themes).