Display leap day anniversaries every year on calendars
Added a new anniversary method to dates which returns a day and
month on which to display leap day anniversaries on non-leap
years.

Added an option in the preferences display tab which allows the
user to specify on which day the anniversary should be displayed.
The options are: 28th Feb, 1st Mar or only on 29th Feb.

Fixes #12511