Allow the user to directly input a Gramps ID in selectors
When a selector is focused, presso Ctrl-O to trigger an input dialog and
insert a Gramps ID directly.

The Gramps ID, if valid, will be used to populate the reference,
bypassing the usual tree model loading mechanism of selection dialogs.

This is a proof-of-concept commit. At a minimum, the feature should be
made available as a button on selector dialogs, as currently the only
way to access it is to know the "magic shortcut" Ctrl-O.