Fix htmldoc image filenames
Filenames of images generated by HtmlDoc did not
contain the crop value. This had the effect that
photos of several people from the same source image
(but with different crops) were overwritten, so all
images of those people showed the same person.

The issue is fixed by adding a crop suffix.