Fix garbage collection issue in ConfigManager
Instances of the python configparser are not removed by the
garbage collector unless some internal objects are deleted first.
This is not an issue on the first run where we load the main
Gramps ini file, but on subsequent runs there are uncollected
objects after loading the view configuration files.

This fix is not ideal because it accesses protected attributes
of the configparser.