Improve AIO build script to run without modifications
The build script required modification before first use to set the full path to the local aio directory. With this change, by invoking the script from the aio directory, that local modification is no longer necessary. Internally the script uses pushd/popd to return to the aio directory.

Second change is to use the `-N` option in wget to avoid downloading a file if the timestamp hasn't changed on the server. This should reduce build time on second and subsequent builds if the source package hasn't changed. It also resolves a lurking error where an updated file on the server would not be used because wget appends a counter to the file names whereas the script was always extracting based on the original filename.

Third change is to avoid the script from asking for interactive input on second and subsequent runs when attempting to unzip dict.zip by passing the `-o` to unzip to allow overwrites.