Report simpler Python version string in About and Error dialogs
A gramps user reported on the discourse forum that information in
the about box was truncated, and he could not expand the box to
view the rest of the text.

A look at the text in question suggested a simple change:
the truncated string is intended to be the Python version which is
typically major.minor.patch version, but the text being displayed
is the output of sys.version() which includes additional information
including build number and compiler used. For the About box this
isn't necessarily useful, so this PR replaces that call
with platform.python_version() which returns just the version and
resolves the truncated string issue. The same change is also made
in the error report dialog for consistency.