Use regular buttons to copy owner information from/to preferences
It's strange to me to use a context menu *anywhere in the window* for
this, and in the space used for the label suggesting how to use it, we
can fit the two buttons instead.

Also, GtkMenu is gone in GTK4, so removing one of its usages will make
such a port easier.