Convert navigator from a Notebook to a Stack
The latter does not need to be homogeneous, so there's no need to fiddle
around with hiding/showing children to auto-resize. It also has a name
for each child, so the link with the combo box can be done automatically
with a binding.

The only downside is that the signal for page changes doesn't tell us
the old page, so we need to track that in order to deactivate it.