Fix fan charts on HiDPI screens
Most DrawingArea-based widgets are fine, but the fan charts use a
backing image surface that is blitted when needed. On HiDPI displays,
this image surface is normal size, and thus everything ends up blurry
when scaled up.

Instead the backing surface should be scaled to match the display, and
also re-created if the display scale changes (e.g., if display settings
change, or the window is moved between displays with different
settings.)