Fix #13261 Installing 5.1.x after 5.2.x will not override config/data locations
In an 5.2.x installations with XDG directory layout for data/config/cache, installing 5.1.x later would create a second config folder in the legacy layout, overriding the existing config.

This change makes it so that if an existing configuration exists for 5.2.x, it will be given preference over the 5.1.x config. If there isn't one, previous logic will continue to be used.