Fix unittest for local timezones
The Unit Test (gramps.plugins.test.exports_test.test_ged) is broken when the
OS is not in UTC time zone. This patch improves the test filter to ignore
dates and times belonging to the last change (CHAN).

Also, the buffer for the previous tokens had to be increased.

Fixes #12902.