Disable rounding glyph positions in cairo document generator
According to the Pango documentation, rounding glyph positions
makes Pango render glyphs so that their positions are integral
in device units. This is useful for rendering text on the screen,
however it does not make sense for documents.

Moreover, with this option enabled, letter spacing can be weird
for some font sizes. It is especially noticeable for sizes like
3-4 pt, some artifacts also appear at bigger sizes. Rendering
large PDF report on small page can produce quite ugly output.