Fix PDF generation failure with LaTeX-based reports
With this change generating PDFs from LaTeX-based reports such as sandclock and other tree reports is fixed.

Gramps set the output directory for lualatex which directs all output from lualatex to that directory. Lualatex looks for a format file when it runs, and if one does not exist it will create one, and saves it to the output directory. However, this directory is not searched by lualatex, so it fails to locate the format file when attempting to generate the PDF file thus leading to the failure reported in the bug report.

The solution used by this PR is to avoid setting the output directory. This lets lualatex choose destination for the format file to one that is searched for by lualatex. However, a side-effect of this is that the PDF was then being sent to the directory Gramps was launched from which isn't suitable. To resolve this, we use the cwd argument that Popen accepts, and set that to the temporary directory so that the report files are generated in the expected location.

Finally, when Gramps was not launched with a console, PDF generation was launching a window and closing it which was not necessary, so the creation flag CREATE_NO_WINDOW was added to Popen, which avoids that.

Fixes #10696