Restore AIO console executable's icon
Starting with Gramps 5.2.0 and the new AIO installer build scripts,
the icon for GrampsAIO-console app changed from the one that has the
console badge to one without, i.e. the icons for both Gramps and
Gramps-console are the same. This commit restores the icon for the
console badge, making it distinguishable once again.

Fixes #13402